Rename options.js to process-options.js
