+++ lib/process-options.js
