+++ lib/shovel.js
@@ -5,7 +5,7 @@ const spawn = child_process.spawn;
 const path = require('path');
 const fs = require('fs-extra');
 
-const processOptions = require('./options');
+const processOptions = require('./process-options');
 
 const downloadGithubRepo = require('./features/download-github-repo');
 const downloadGist = require('./features/download-gist');
